added exception handling on roottonumpy and changed its name to traducers.py
